pyproject.toml
[tool.poetry]
name = "athena"
version = "0.1.0"
description = ""
authors = ["Invictus808 <contact@invictus808.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
uvicorn = "^0.30.6"
pypdf = "^4.3.0"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
